requirements.txt
# Original LemonFox dependencies
requests>=2.28.0
python-dotenv>=0.21.0
sounddevice>=0.4.5
webrtcvad>=2.0.10
numpy>=1.20.0
scipy>=1.7.0

# Additional voice module dependencies
pynput>=1.7.6
pyautogui>=0.9.53
pyperclip>=1.8.2
pystray>=0.19.4
Pillow>=9.0.0

# Platform-specific dependencies (install manually based on platform)
Windows: pip install pywin32
macOS: pip install pyobjc-framework-Cocoa
Linux: sudo apt-get install python3-xlib xdotool